requirements.txt
streamlit
pandas
pyodbc
python-dotenv
chromadb
sentence-transformers
ollama
sqlglot
vanna
ploty
IPython
sqlalchemy
